[ocp] Add OCP tests for file bindy ftp quickstart
